Ignore this because of a bug in GDB where it classifies any type with a member function as a class.
Even if the debug information classifies the type as a struct, GCC renders it
as a class if it contains a member function. Clang recently started emitting
implicit members for C++ types, including the implicit copy constructor in
structs. This caused the failure but only due to this bug in GCC. More recent
GCC versions do not exhibit this bug.


git-svn-id: http://llvm.org/svn/llvm-project/clang-tests/trunk@167384 91177308-0d34-0410-b5e6-96231b3b80d8